Plan 9 from Bell Labs 2008-01-20

+++ sys/src/cmd/ndb/dnserver.c
@@ -165,7 +165,7 @@ dnserver(DNSmsg *reqp, DNSmsg *repp, Request *req, uchar *srcip, int rcode)
 static RR*
 doextquery(DNSmsg *mp, Request *req, int recurse)
 {
-	int type;
+	ushort type;
 	char *name;
 	RR *rp, *neg;
